type references also expand type arguments
