+++ src/main/java/info/kfgodel/bean2bean/dsl/impl/scopes/ExplicitScopeWithDefinedSourceImpl.java
@@ -5,6 +5,9 @@ import info.kfgodel.bean2bean.core.api.registry.DomainVector;
 import info.kfgodel.bean2bean.dsl.api.scopes.ExplicitScopeWithDefinedSource;
 import info.kfgodel.bean2bean.dsl.api.scopes.ScopedRegistrationDsl;
 import info.kfgodel.bean2bean.dsl.impl.ConfigureDslImpl;
+import info.kfgodel.bean2bean.other.references.TypeRef;
+
+import java.lang.reflect.Type;
 
 /**
  * Default implementation
@@ -24,10 +27,20 @@ public class ExplicitScopeWithDefinedSourceImpl<I> implements ExplicitScopeWithD
 
   @Override
   public <O> ScopedRegistrationDsl<I, O> andProduce(Class<O> targetType) {
+    return this.andProduce((Type)targetType);
+  }
+
+  @Override
+  public <O> ScopedRegistrationDsl<I, O> andProduce(TypeRef<O> targetTypeRef) {
+    return andProduce(targetTypeRef.getReference());
+  }
+
+  private <O> ScopedRegistrationDsl<I, O> andProduce(Type targetType) {
     Domain targetDomain = parentDsl.calculateDomainFor(targetType);
     DomainVector conversionVector = DomainVector.create(inputDomain, targetDomain);
     return ScopedRegistrationDslImpl.<I,O>create(conversionVector, getConfigureDsl());
   }
+
 
   private ConfigureDslImpl getConfigureDsl() {
     return this.parentDsl.getParentDsl();
